huaweiwei #9 优化引擎代码
huaweiwei #9 类型安全: 修复了 int/size_t/uint32_t 混用问题
资源管理: 完善了 libuv handles 的生命周期管理
错误处理: 增强了内存分配失败的处理
功能完整性: 实现了 BSON 嵌套表的完整支持
跨平台兼容: 添加了字节序转换

+++ ge/src/server.h
@@ -18,8 +18,8 @@
 typedef struct {
     uv_tcp_t handle;
     char buffer[CLIENT_BUFFER_SIZE];
-    int buffer_used;             // Bytes used in receive buffer
-    int expected_length;         // Expected message length (from header)
+    size_t buffer_used;          // Bytes used in receive buffer
+    size_t expected_length;      // Expected message length (from header)
     struct server_context* server;
     int array_index;             // Index in server's client array (for O(1) removal)
     uint32_t magic;              // Magic number for validity check
